Move test files + make stats it's own sub packagE
